requirements.txt
atlassian-python-api
boto3